requirements/production.txt
Flask
Flask-PyMongo
pydicom
pillow
numpy